chore(lineage): code refactor
